Test all PDAL filters with their default settings
